Wrote connections to firebase along with handling of changes to water pump state.
