--- controllers/service-account.json
@@ -1,12 +1,0 @@
-{
-  "type": "service_account",
-  "project_id": "adoptaplant-1b4e1",
-  "private_key_id": "3cf3d682c431c28b2b00d01000fe318066992cdd",
-  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQDVh3ntS5IOPfUt\nYoyVPB9U0btg9Gv3LnO9k94HkHoF9LsS9INYqHxfn7+lhgfC0GQEuFUd/aiAXLaa\nv+zCDdCXyfAwTBOEBfIY+5GphMBlxgJogt+Ekmsse+tMcaT2lZYxdyjK7FCJym5w\nKOLoMkrfP7o61t+NNJ+ok5xkQMu2+JXGiHTuiBe8JNQ8+EYtoUfZ4nypIDvjI+Pi\nh7/RgIViyEfiy0ktywb8gJapXqWA87bDrZGXgXY76c6f05V0l4xaQxKPHfqKB4Tr\nt3xNQokWKAXGKW/QTSHqJl9zZ8VagMLnwAZ67MZbHJQlBnsiwughSO68dkAvsVm8\naTST+WzRAgMBAAECggEACiL6zkw1RvjOVVUQKxt7ZJ7qQCGKUzsrd8H3xdsoInU/\nlkLw+KSoTC/SEhez8Z3L9GS9zFqDhR3MS2S2EXDvc0KSZPCPnUeyT65NosMFu9nJ\nDntmr1f2FNqzImIOEeirBHZ4Lh3xetZsehBEKVVjNDLwcnpLNYcHLKJenRcYX1Cv\n32YPzuAa1FYsEDA5V1tOCdW6wQz1+ASgdj4wTf9sw4xM/tYcdayWoxmsc62UPOgZ\nFfRn3qD8lGfv5PnAR4UkbWYn9ZlqFoYaFg7/fqYrGtC0Q5lidTvFHrLK4WHSUmmv\nvbJFPLCL08gIOsR9wgtjzWjBO9aLvKNIPFr2PCXZQwKBgQDqMQiGt8spXnCK28RO\n0e65ZaAW+80oo9/zxj4rMi0+SMaYniHw7L0uuCkTBfyK4fBX8qj28XoDZ60vmpEP\nHWlSRQaWc2ieE1vgtZth3YNP5ePBUhFCskxRRqzEYHNJzO9nlQXd9byVC15qV5fV\nz7jPojsoJgPOyN1AAxEE3PaLTwKBgQDpad324uJB8bhCG+jkpUZnB0sAlN5P7vm7\nn7Xc2kNZEbfNs7yIpCXVHGULsZaiu/MXqef/UgrcImDGKugZ0BAE6QPtEED+8Fpu\nevKElbUtc18ZaFfOACgJFpIJplK8dLsheScpt3fR1baQl4uYEtv5aM4P77rY0V8P\nwZyLH6f93wKBgDoW74Z7EGmygzfuS0NBz3UtQAhzbPAnteVBxynvyEkqxBgXIVQh\nZMux4n5Vfc71PCiUiFS4kQtHLC4iqNiW8wj1Fi07czfClYenhGNCDtBCtbu+H4nr\nsAJ5BWGwiuhDbR7EWDA2RBC0QTcfgwoF/FGEbCgJReiE6lx2ZkJOjNkBAoGADgom\nJcGD5n+Jm8BeA9LFJ1FS+5f4v0oaj2SN5djjECp2qdLT1kw5UlZA5bEq5GL1vkik\nAcGHkbInrAPMrjDLpd/2Mn41yp/ecKLIj50JptA3uimCjouH3PQUmKv+cgN7dgn5\nhxLYDDEUqCHVbbJEtP3O1zqOPJhrZWE570xaKHUCgYANrwDGvLfGEFrDsi2AjMoo\nnUzsJEZ2/Zd38dPsn8dSLpX9l9/6voqcSxcNe4X6E4vUQRYmZlJzdr8wSDBx5U0j\nSTZsavuW/+EN+ieMNpMELOrgiZMdQBMUBJ+TNHopA2JCNFnQG0jrZkyAGQ5ZnhWz\nD/rujYQ2MbyHzDU9u6la1w==\n-----END PRIVATE KEY-----\n",
-  "client_email": "firebase-adminsdk-yhzr9@adoptaplant-1b4e1.iam.gserviceaccount.com",
-  "client_id": "101731581315769911125",
-  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-  "token_uri": "https://oauth2.googleapis.com/token",
-  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-yhzr9%40adoptaplant-1b4e1.iam.gserviceaccount.com"
-}
